// backend/src/venues/enums/venue.enums.ts
export enum VenueStatus{
    PENDING_APPROVAL = 'PENDING_APPROVAL',
    APPROVED = 'APPROVED',  
    REJECTED = 'REJECTED'
}

export enum WeekDays{
    MONDAY = "MONDAY",
    TUESDAY = "TUESDAY",
    WEDNESDAY = "WEDNESDAY",
    THURSDAY = "THURSDAY",
    FRIDAY = "FRIDAY",
    SATURDAY = "SATURDAY",
    SUNDAY = "SUNDAY"
}

export enum Districts {
    ALAPPUZHA = "Alappuzha",
    ERNAKULAM = "Ernakulam",
    IDUKKI = "Idukki",
    KANNUR = "Kannur",
    KASARAGOD = "Kasaragod",
    KOLLAM = "Kollam",
    KOTTAYAM = "Kottayam",
    KOZHIKODE = "Kozhikode",
    MALAPPURAM = "Malappuram",
    PALAKKAD = "Palakkad",
    PATHANAMTHITTA = "Pathanamthitta",
    THIRUVANANTHAPURAM = "Thiruvananthapuram",
    THRISSUR = "Thrissur",
    WAYANAD = "Wayanad"
}

export enum SlotStatus{
    AVAILABLE = 'AVAILABLE',
    HELD = 'HELD',
    BOOKED = 'BOOKED',
};